update anisotropyCheck and add VDF_plot
Spent a while messing up the anisotropyCheck plot, now have it using
realistic data but still the old gaussian system.. I’ll just have to
increase the number of points on those. Also added the VDF_plot
routine, which is just the first half of the anisotropy finding routine
for now.