anisotropyAveraging update
tried to tweak this to run on the mac smoothly but ran into problems
with the uncertainyMaps variables. I never fixed those, so they don’t
have a map variable embedded. It is time to finally create a subprocess
to pull the data and create a map, that will be done on the laptop