progress with dHT as a sliding window calculation. Next step is to isolate switchback boundaries and get a measure of switchback expansion
Made a whole mess of the anisotropyFromMath.jy code when trying to kludge in the extra peak measurements. It needs to be rewritten as a subprocess that I can feed different VDFs into, because I am just getting errors by iterating at different rates. It will be slower, but I think it is necessary.