Added an option to attempt to rewrite the file multiple times. I have found I often get broken cdf files when some process on the Windows side interrupts the write process by Autoplot. This checks the Epoch value of the file after writing, and rewrites if the Epoch values are incorrect.
Also added the ability to turn off the v0 file generation, which would happen if there was not efield data found. That idea was a workaround for when I wanted to look at early data before it was fully processed, but slows down general processing if I am scanning through old files to replace bad writes.